feat: upgrade to 64-bit types and expand documentation
Changes API to use 64-bit types throughout:
- Changed flen() return type from int to int64_t for large file support
- Updated heapinfo() parameters to uint64_t for 64-bit address spaces
- Changed parameter array from int to int64_t to support 64-bit values
- Updated start_clock from uint32_t to uint64_t to prevent overflow

Also improved README with more detailed explanation of Zero Board
Computer architecture and semihosting use cases, and updated
documentation URL to johnwbyrd.github.io.

+++ include/zbc_backend_ansi.h
@@ -144,9 +144,8 @@ typedef struct zbc_ansi_state_s {
 
     /*--- Internal: other state ---*/
     int last_errno;
-    /* TODO: uint32_t will overflow after ~497 days of centiseconds.
-     * Consider using uint64_t or detecting/handling overflow. */
-    uint32_t start_clock;              /**< Stored as uint32_t to avoid time.h */
+    /* Clock accumulator using 64-bit time */
+    uint64_t start_clock;              /**< Stored as uint64_t to avoid time.h */
     char path_buf[ZBC_ANSI_PATH_BUF_MAX];
     int initialized;
 } zbc_ansi_state_t;
@@ -232,9 +231,8 @@ typedef struct zbc_ansi_insecure_state_s {
 
     /*--- Internal: other state ---*/
     int last_errno;
-    /* TODO: uint32_t will overflow after ~497 days of centiseconds.
-     * Consider using uint64_t or detecting/handling overflow. */
-    uint32_t start_clock;              /**< Stored as uint32_t to avoid time.h */
+    /* Clock accumulator using 64-bit time */
+    uint64_t start_clock;              /**< Stored as uint64_t to avoid time.h */
     char path_buf[ZBC_ANSI_PATH_BUF_MAX];  /**< For null-terminating paths */
     int initialized;
 } zbc_ansi_insecure_state_t;
